some more warning fixes
..could actually minimally influence the emulation, so a separate commit

git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3929 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/adpcm.h
@@ -39,7 +39,7 @@ int ADPCM_playing(int num);
 struct OKIM6295interface
 {
 	int num;                  		/* total number of chips */
-	int frequency[MAX_OKIM6295];	/* playback frequency */
+	float frequency[MAX_OKIM6295];	/* playback frequency */
 	int region[MAX_OKIM6295];		/* memory region where the sample ROM lives */
 	int mixing_level[MAX_OKIM6295];	/* master volume */
 };
